Validation for delete operation for blogs added

+++ part4/bloglist/backend/controllers/blogs.js
@@ -12,6 +12,10 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const decodedToken = jwt.decode(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'missing token' })
+  }
+
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'invalid token' })
   }
@@ -33,6 +37,21 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.decode(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'missing token' })
+  }
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
